Refactor and Feat: User can remove companies from database.
Refactor: Created new function getStock which call yahoo-finance API to get stock data.
This is done becuase stock data are not stored in the database.
When the user adds or deletes companies, the API will be called again, so a function is created for that to avoid code dublications.
Feat: When the user clicks the delete button, the company gets removed from the database.