Refacot: getStock function updated and 'newCompany' event now working.
getStock function given new parameter isCompaniesModified.
This paramter decides to send results to all clients or just one client.
This is done because this function is used when page loads and when a new company is added.
It is needed to send data to all clients only when a new company is added.
newComany event is added. It is called when the user adds a new company.
It saves the new company to the database and calls getStock function.