Bug: Fixed bug which occured when deleting a company.
When the client tries to remove a company, the wait div appears and nothing happens.
This happened because the condition to to emit to the client is always false.
A status parameter is added to getStock function to indicate whether its called because of removing company, adding one, or simply the page is loaded.